Agregando katex

+++ app/Livewire/Question/QuestionList.php
@@ -13,12 +13,22 @@ class QuestionList extends Component
     {
         abort_if(!auth()->user()->is_admin, Response::HTTP_FORBIDDEN, 403);
 
+        $user = auth()->user();
+
+        // Solo permitir eliminar si la pregunta pertenece al usuario actual
+        if ($question->user_id !== $user?->id) {
+            abort(403, 'No autorizado');
+        }
         $question->delete();
     }
 
     public function render(): View
     {
-        $questions = Question::latest()->paginate();
+        // $questions = Question::latest()->paginate();
+        $user = auth()->user();
+        $questions = Question::where('user_id', $user?->id)
+            ->latest()
+            ->paginate();
 
         return view('livewire.question.question-list', compact('questions'));
     }
